# Accelerate point cloud conversion to ROS messsge