--- a/carla_ros_bridge/src/carla_ros_bridge/sensor.py
+++ b/carla_ros_bridge/src/carla_ros_bridge/sensor.py
@@ -13,7 +13,6 @@ from __future__ import print_function
 
 import ctypes
 import os
-
 try:
     import queue
 except ImportError:
@@ -45,6 +44,7 @@ _DATATYPES[PointField.FLOAT64] = ('d', 8)
 
 
 class Sensor(Actor):
+    
     """
     Actor implementation details for sensors
     """
@@ -176,7 +176,7 @@ class Sensor(Actor):
         if self.synchronous_mode:
             if self.sensor_tick_time:
                 self.next_data_expected_time = carla_sensor_data.timestamp + \
-                                               float(self.sensor_tick_time)
+                    float(self.sensor_tick_time)
             self.queue.put(carla_sensor_data)
         else:
             self.publish_tf(trans.carla_transform_to_ros_pose(
@@ -208,8 +208,8 @@ class Sensor(Actor):
                 if carla_sensor_data.frame != frame:
                     self.node.logwarn("{}({}): Received event for frame {}"
                                       " (expected {}). Process it anyways.".format(
-                        self.__class__.__name__, self.get_id(),
-                        carla_sensor_data.frame, frame))
+                                          self.__class__.__name__, self.get_id(),
+                                          carla_sensor_data.frame, frame))
                 self.node.logdebug("{}({}): process {}".format(
                     self.__class__.__name__, self.get_id(), frame))
                 self.publish_tf(trans.carla_transform_to_ros_pose(
@@ -220,9 +220,9 @@ class Sensor(Actor):
 
     def _update_synchronous_sensor(self, frame, timestamp):
         while not self.next_data_expected_time or \
-                (not self.queue.empty() or
-                 self.next_data_expected_time and
-                 self.next_data_expected_time < timestamp):
+            (not self.queue.empty() or
+             self.next_data_expected_time and
+             self.next_data_expected_time < timestamp):
             while True:
                 try:
                     carla_sensor_data = self.queue.get(timeout=1.0)